requirements.txt
streamlit
google-genai
pymupdf
pandas
openpyxl
pillow
python-dotenv